add hook for random countries
