+++ flags-fe/src/components/Quiz/Quiz.jsx
@@ -1,14 +1,16 @@
 import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
+import { useRandomSelector } from "../../hooks/useRandomSelector";
 
 import Question from "../Question/Question";
 import FlagsBoard from "../FlagsBoard/FlagsBoard";
-
-import styles from "./Quiz.module.css"
 import SetupQuiz from "../SetupQuiz/SetupQuiz";
 
+import styles from "./Quiz.module.css";
 
 export default function Quiz() {
+
+    const {selectRandomCountries} = useRandomSelector();
 
     const [gameList, setGameList] = useState([]);
     const [options, setOptions] = useState([]);
@@ -39,20 +41,6 @@ export default function Quiz() {
         setIsStarted(true);
     }
 
-    const selectRandomCountries = (list, count) => {
-        const selected = [];
-        const countryIndices = [];
-        let i = 0;
-        while (i < count && i < list.length) {
-            const randomIndex = Math.floor(Math.random() * list.length);
-            if (!countryIndices.includes(randomIndex)) {
-                countryIndices.push(randomIndex);
-                selected.push(list[randomIndex]);
-                i++;
-            }
-        }
-        return selected;
-    };
 
     const answerHandler = (event) => {
         const choice = event.target.textContent;
